testing/web-platform/tests/old-tests/submission/Infraware/Session_History/contents/Session_History/combination_history_006.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Session History</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>combination_history_006(After calling of go method, check length.)</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<script>
		
		function onPopstateEvent(e) {
			if (e.state == 1 && last == window.history.length) {
				test(function() {
					assert_true(true, "combination test is failed.");
				});
			} else {
				test(function() {
					assert_unreached("combination test is failed.");
				});
			}
		}

		if (window.addEventListener) {
            window.addEventListener('popstate', onPopstateEvent, false);
        }
        else if (window.attachEvent) {
            window.attachEvent("onpopstate", onPopstateEvent);
        }
        else {
            test(function() {
				assert_unreached("DOM event subscribtion is not supported.");
			});
        }

		var last;
		try {
			window.history.pushState(1, document.title, '?x=1');
			window.history.pushState(2, document.title, '?x=2');
		} catch (e) {
			test(function() {
				assert_unreached("call of pushState method is failed.");
			});
		}
		last = window.history.length;

		window.history.go(-1);
	</script>

 </body>
</html>
